Move and rename everest_client
